+++ src/deluge/model/song/song.cpp
@@ -5263,8 +5263,13 @@ bool Song::deletePendingOverdubs(Output* onlyWithOutput, int32_t* originalClipIn
 
 			anyDeleted = true;
 		}
+		// this isn't actually deleting, but it is clearing a pending overdub which is what matters for the caller
+		else if (clip->armState == ArmState::ON_TO_RECORD) {
+			clip->armState = ArmState::OFF;
+			anyDeleted = true;
+		}
 	}
-
+	D_PRINTLN("Deleted pending overdubs");
 	return anyDeleted;
 }
 
@@ -5313,28 +5318,33 @@ Clip* Song::getClipWithOutputAboutToBeginLinearRecording(Output* output) {
 }
 
 Clip* Song::createPendingNextOverdubBelowClip(Clip* clip, int32_t clipIndex, OverDubType newOverdubNature) {
-
+	Clip* newClip = clip; // we're returning the clip to be scheduled, default to this one
 	// No automatic overdubs are allowed during soloing, cos that's just too complicated
 	if (anyClipsSoloing) {
 		return NULL;
 	}
+	// if we're in rows or the clip can't support in place overdub then use the traditional deluge cloning looper
+	if (sessionLayout == SessionLayoutType::SessionLayoutTypeRows || clip->shouldCloneForOverdubs()) {
+		char modelStackMemory[MODEL_STACK_MAX_SIZE];
+		ModelStack* modelStack = setupModelStackWithSong(modelStackMemory, this);
 
-	char modelStackMemory[MODEL_STACK_MAX_SIZE];
-	ModelStack* modelStack = setupModelStackWithSong(modelStackMemory, this);
+		ModelStackWithTimelineCounter* modelStackWithTimelineCounter = modelStack->addTimelineCounter(clip);
 
-	ModelStackWithTimelineCounter* modelStackWithTimelineCounter = modelStack->addTimelineCounter(clip);
+		newClip = clip->cloneAsNewOverdub(modelStackWithTimelineCounter, newOverdubNature);
 
-	Clip* newClip = clip->cloneAsNewOverdub(modelStackWithTimelineCounter, newOverdubNature);
+		if (newClip && newClip != clip) {
+			newClip->overdubNature = newOverdubNature;
+			sessionClips.insertClipAtIndex(newClip, clipIndex);
+			if (clipIndex != songViewYScroll) {
+				songViewYScroll++;
+			}
 
-	if (newClip) {
-		newClip->overdubNature = newOverdubNature;
-		sessionClips.insertClipAtIndex(newClip, clipIndex);
-		if (clipIndex != songViewYScroll) {
-			songViewYScroll++;
+			// use root UI in case this is called from performance view
+			sessionView.requestRendering(getRootUI());
 		}
-
-		// use root UI in case this is called from performance view
-		sessionView.requestRendering(getRootUI());
+	}
+	else {
+		clip->setupOverdubInPlace(newOverdubNature);
 	}
 
 	return newClip;